requirements.txt
# Web Framework
flask>=2.0.0

# Database
psycopg2-binary>=2.9.0
supabase>=0.0.1

# LLM Clients
openai>=1.0.0
ollama>=0.1.0

# Slack Integration
slack-bolt>=1.16.0

# Environment Variables
python-dotenv>=0.19.0

# Utilities
tabulate>=0.8.0
requests
pprintpp

# Replit (optional, can be removed for non-Replit deployments)
replit>=3.0.0
